This removes TwinComp and SecondTrack. That is
   because the functionality is repoduced by FixedGroup
   and it seemed a mess. Additionally there are some
   fixes for some old models

+++ System/process/objectRegister.cxx
@@ -3,7 +3,7 @@
  
  * File:   process/objectRegister.cxx
  *
- * Copyright (c) 2004-2018 by Stuart Ansell
+ * Copyright (c) 2004-2019 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -53,8 +53,6 @@
 #include "HeadRule.h"
 #include "LinkUnit.h"
 #include "FixedComp.h"
-#include "SecondTrack.h"
-#include "TwinComp.h"
 #include "FixedGroup.h"
 #include "ContainedComp.h"
 #include "SpaceCut.h"
@@ -146,7 +144,7 @@ objectRegister::addObject(const std::string& Name,
 
   
 int
-objectRegister::cell(const std::string& Name,const int size)
+objectRegister::cell(const std::string& Name,const size_t size)
   /*!
     Add a component and get a new cell number 
     \param Name :: Name of the unit
